refactor : BlackCustomer 를 다시 분리
