--- src/main/java/org/prgrms/kdtspringdemo/domain/customer/repository/BlackListRepository.java
@@ -1,9 +1,0 @@
-package org.prgrms.kdtspringdemo.domain.customer.repository;
-
-import org.prgrms.kdtspringdemo.domain.customer.model.BlackCustomer;
-
-import java.util.List;
-
-public interface BlackListRepository {
-    List<BlackCustomer> findAllBlackCustomers();
-}
